fix for DERBY-298.  committing change for ·¹ystein Gr±±±vlen
 The attached patch fixes the bug by setting the logEnd after recovery to the beginning of the new empty log file instead of the end of the previous file.
The patch contains changes to the following files:

M java/engine/org/apache/derby/impl/store/raw/log/FileLogger.java
        - At the end of the redo scan, if the scan stopped in a file succeeding the file of the last log record, update logEnd to this position.
        - Change assert to allow logEnd to be in a newer file than that of the last log record.
         
M java/engine/org/apache/derby/impl/store/raw/log/Scan.java
        - Introduced new variable newFileStart which will only have a valid LogInstant value when the scan is at the header of the file.
        - When a new file is entered, set newFileStart to the first possible LogInstant of this file (end of header).
        - When a log record is encountered, set newFileStart to INVALID_LOG_INSTANT.
        - Changed getLogRecordEnd() to return newFileStart if that is valid (i.e., scan is at the start of a file)
        - Removed comment about not starting to write to the new empty log file, since that is not true anymore.

A java/testing/org/apache/derbyTesting/functionTests/tests/store/RecoveryAfterBackup_app.properties
         - Test properties

M java/testing/org/apache/derbyTesting/functionTests/tests/store/copyfiles.ant
         - Added new property files

A java/testing/org/apache/derbyTesting/functionTests/tests/store/RecoveryAfterBackupSetup_app.properties
        - Test properties.
        - useextdirs=true needed so the backup is placed somewhere the next test can find it.

A java/testing/org/apache/derbyTesting/functionTests/tests/store/RecoveryAfterBackup.java
        - Test that is supposed to be run after RecoveryAfterBackupSetup.java.
        - Does recovery, updates the database, shutdowns the database, and does roll-forward restore.
        - Checks that updates made after recovery is reflected in the database after roll-forward restore.

A java/testing/org/apache/derbyTesting/functionTests/tests/store/RecoveryAfterBackupSetup.java
        - Test that does the preparation for the RecoveryAfterBackup test.
        - Inserts a few records, makes a backup, and stops without shutting down.

M java/testing/org/apache/derbyTesting/functionTests/harness/RunTest.java
        - For tests where the database is not deleted at the end of the test, do not delete the external directories either.
        - This is necessary to be able to access the backup in suceeding tests.

A java/testing/org/apache/derbyTesting/functionTests/master/RecoveryAfterBackupSetup.out
        - Test output

A java/testing/org/apache/derbyTesting/functionTests/master/RecoveryAfterBackup.out
        - Test output

MM java/testing/org/apache/derbyTesting/functionTests/suites/storerecovery.runall
        - Added tests to storerecovery suite.
        - Changed property eol-style.

 The recently attached patch (derby-298a.diff) addresses Suresh's
review comments. The only major change from the previous patch is in
java/engine/org/apache/derby/impl/store/raw/log/Scan.java. The changes
to this file compared to the current head of trunk are:
 
   - When a new log file is entered, check that the header of this
     file refers to the end of the last log record of the previous log
     file. If not, stop the scan.
   - If the header was consistent, update knowGoodLogEnd to the first
     possible LogInstant of this file (end of header).
   - close() no longer reset knownGoodLogEnd since it is needed by
     FileLogger after the scan is closed.
   - Changed comment of getLogRecordEnd() to reflect that it can be
     used after the scan is closed, and that it at that time may
     return the start of an empty log file.
   - Removed comment about not starting to write to the new empty log
     file, since that is not true anymore.

In addition, the property files for the tests have been updated so
they are run without the security manager. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@367352 13f79535-47bb-0310-9956-ffa450edef68

+++ java/testing/org/apache/derbyTesting/functionTests/harness/RunTest.java
@@ -1852,15 +1852,13 @@ clp.list(System.out);
 			        //System.out.println(logfile.getPath());
 			        //status = logfile.delete();
 		        }
-            }
 
-			//delete the directories where external input/output files were created
-			if(extInDir!=null)
-				deleteFile(extInDir);
-			if(extOutDir!=null)
-				deleteFile(extOutDir);
-			if(extInDir!=null)
-				deleteFile(extInOutDir);
+                // delete the directories where external input/output files 
+                // were created
+                if (extInDir!=null) deleteFile(extInDir);
+                if (extOutDir!=null) deleteFile(extOutDir);
+                if (extInDir!=null) deleteFile(extInOutDir);
+            }
         }
 	    // reset for next test
 	    // the next line is a bug fix to get cleanup working correctly when
